requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.1.4
django-crispy-forms==2.0
django-filter==22.1
django-tinymce==3.5.0
Pillow==9.4.0
sqlparse==0.4.3
